+++ src/app/cli/src/commands/system_info_command.rs
@@ -6,28 +6,39 @@
 // As of the Change Date specified in that file, in accordance with
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
-
 use std::sync::Arc;
 
+use container_runtime::ContainerRuntime;
 use internal_error::*;
+use serde_json::Value;
 
 use super::{CLIError, Command};
 use crate::output::*;
+use crate::{WorkspaceService, WorkspaceVersion};
 
 ///////////////////////////////////////////////////////////////////////////////
 
 pub struct SystemInfoCommand {
     output_config: Arc<OutputConfig>,
     output_format: Option<String>,
+    workspace_svc: Arc<WorkspaceService>,
+    container_runtime: Arc<ContainerRuntime>,
 }
 
 impl SystemInfoCommand {
-    pub fn new<S>(output_config: Arc<OutputConfig>, output_format: Option<S>) -> Self
+    pub fn new<S>(
+        output_config: Arc<OutputConfig>,
+        container_runtime: Arc<ContainerRuntime>,
+        workspace_svc: Arc<WorkspaceService>,
+        output_format: Option<S>,
+    ) -> Self
     where
         S: Into<String>,
     {
         Self {
             output_config,
+            container_runtime,
+            workspace_svc,
             output_format: output_format.map(|s| s.into()),
         }
     }
@@ -41,7 +52,7 @@ impl Command for SystemInfoCommand {
 
     async fn run(&mut self) -> Result<(), CLIError> {
         write_output(
-            SystemInfo::collect(),
+            SystemInfo::collect(self.container_runtime.clone(), self.workspace_svc.clone()).await,
             &self.output_config,
             self.output_format.as_ref(),
         )?;
@@ -117,15 +128,24 @@ fn write_output<T: serde::Serialize>(
 #[serde(rename_all = "camelCase")]
 pub struct SystemInfo {
     pub build: BuildInfo,
+    pub workspace: WorkspaceInfo,
+    pub container_runtime: ContainerRuntimeInfo,
 }
 
 impl SystemInfo {
-    pub fn collect() -> Self {
+    pub async fn collect(
+        container_runtime_svc: Arc<ContainerRuntime>,
+        workspace_svc: Arc<WorkspaceService>,
+    ) -> Self {
         Self {
             build: BuildInfo::collect(),
+            workspace: WorkspaceInfo::collect(workspace_svc),
+            container_runtime: ContainerRuntimeInfo::collect(container_runtime_svc).await,
         }
     }
 }
+
+///////////////////////////////////////////////////////////////////////////////
 
 #[derive(Debug, serde::Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -164,3 +184,74 @@ impl BuildInfo {
         }
     }
 }
+
+///////////////////////////////////////////////////////////////////////////////
+
+#[derive(Debug, serde::Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct WorkspaceInfo {
+    pub version: u32,
+    pub root_dir: String,
+}
+
+impl WorkspaceInfo {
+    pub fn collect(workspace_svc: Arc<WorkspaceService>) -> Self {
+        if !workspace_svc.is_in_workspace() {
+            return Self {
+                version: 0,
+                root_dir: "".to_string(),
+            };
+        }
+
+        let root_dir = match workspace_svc.layout() {
+            Some(wl) => wl.root_dir.to_str().map(String::from).unwrap(),
+            None => "".to_string(),
+        };
+
+        let workspace_version = workspace_svc
+            .workspace_version()
+            .unwrap_or_default()
+            .unwrap_or(WorkspaceVersion::Unknown(0));
+
+        Self {
+            version: workspace_version.into(),
+            root_dir,
+        }
+    }
+}
+
+///////////////////////////////////////////////////////////////////////////////
+
+#[derive(Debug, serde::Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ContainerRuntimeInfo {
+    pub info: Value,
+}
+
+impl ContainerRuntimeInfo {
+    pub async fn collect(container_runtime: Arc<ContainerRuntime>) -> Self {
+        let container_info_output = match container_runtime.info().output().await {
+            Ok(container_info) => {
+                if container_info.status.success() {
+                    match serde_json::from_slice(&container_info.stdout) {
+                        Ok(res) => res,
+                        Err(_) => {
+                            serde_json::json!({"error": "Unable to parse container runtime info result"})
+                        }
+                    }
+                } else {
+                    serde_json::json!({"error": "Unable to run container runtime info command"})
+                }
+            }
+            Err(_) => {
+                serde_json::json!({"error": "Unable to fetch container runtime info"})
+            }
+        };
+
+        Self {
+            info: container_info_output,
+        }
+    }
+}
+
+///////////////////////////////////////////////////////////////////////////////